pyproject.toml
[project]
name = "modern-elt-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "dagster>=1.13.2",
    "dagster-dbt>=0.29.2",
    "dagster-postgres>=0.29.2",
    "dagster-webserver>=1.13.2",
    "dbt-postgres>=1.10.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.49",
]

[tool.dagster]
module_name = "orchestration.definitions"
